refactor: extract tool-runner + forensic-gate routes into routes/tools.ts
Pure structural move out of createApp; no handler logic changed. Graduated onto
RouteContext (all shared with the drop-folder code that stays in createApp):
runToolAndIngest + reloadCustomTools (stable hoisted methods) and the
liveToolConfigs + customTools live accessors.
